자료형 바꿔줌... TimeStamp 쓰면 안되네... Amount 관련 컬럼들은 int할지 float할지 고민중

+++ ReGoProject/src/main/java/com/smhrd/entity/r_recipe_msg.java
@@ -29,5 +29,5 @@ public class r_recipe_msg {
 
     // 식재료 투입량
 	@Column
-    private Integer msgAmount;
+    private float msgAmount;
 }
